pyproject.toml
[project]
name = "pyatorochkaapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.131.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
